// src/app/coins/types.ts
export interface ICurrencyData {
    code: string;
    codein: string;
    name: string;
    high: string;
    low: string;
    varBid: string;
    pctChange: string;
    bid: string;
    ask: string;
    timestamp: string;
    create_date: string;
}
  
export interface ICurrency {
    [currencyPair: string]: ICurrencyData;
}